misc,tests: Wipe cache and attempt MacOS rebuild in Daily tests
As the environment is not fixed, there can be instances where the
cached MacOS build becomes incompatible with the currenty MacOS build
environment (for example, when a dependency version changes, the gem5
cached code built with the older version can't be used to build gem5
in an environemnt with the newer dependency version).

This is fixed by catching the compilation error and wiping the cached
content then trying again, building from scatch. If successful then
this newer version will replace the older cache.